requirements.txt
pytest
sqlalchemy
psycopg2-binary
alembic
python-dotenv
pybind11
PyQt6
pytest-qt
